after cleanup

+++ GameServerCore/Domain/GameObjects/IObjAiBase.cs
@@ -69,7 +69,7 @@ namespace GameServerCore.Domain.GameObjects
         /// <summary>
         /// A unit's acquisition range is the distance at which they will automatically acquire attack targets, referred to as 'auto-attacking'.
         /// </summary>
-        float AcquisitionRange { get; set; }
+        int AcquisitionRange { get; set; }
         /// <summary>
         /// AIScript should set this to true if it can handle incoming calls for help.
         /// To have callers populate a list of units attacking allies.
